rename the rename tag command to just rename
