+++ gitblend/cli.py
@@ -2,8 +2,7 @@ import argparse
 import sys
 
 from gitblend.commands import update_all, version
-from gitblend.commands.tags import (create_tag, delete_tag, list_tags,
-                                    rename_tag)
+from gitblend.commands.tags import create_tag, delete_tag, list_tags, rename
 
 
 def add_create_tag_command(subparsers):
@@ -47,7 +46,7 @@ def add_rename_tag_command(subparsers):
         "old_tag", type=str, help="The name of the tag to rename"
     )
     rename_tag_parser.add_argument("new_tag", type=str, help="The new name for the tag")
-    rename_tag_parser.set_defaults(func=rename_tag.run)
+    rename_tag_parser.set_defaults(func=rename.run)
 
 
 def add_update_all_command(subparsers):
