+++ tests/test_cli.py
@@ -1,5 +1,5 @@
 import unittest
-from unittest.mock import MagicMock, patch
+from unittest.mock import patch
 
 from gitblend.cli import main
 
@@ -31,7 +31,7 @@ class TestCLI(unittest.TestCase):
         main()
         mock_list_tags.assert_called_once()
 
-    @patch("gitblend.commands.tags.rename_tag.run", side_effect=mock_run)
+    @patch("gitblend.commands.tags.rename.run", side_effect=mock_run)
     @patch("sys.argv", ["gitblend", "rename-tag", "v1.0", "v1.1"])
     def test_rename_tag_command(self, mock_rename_tag):
         main()
